Twitter: The media entities array of a status message should never be NULL
In the response for a status message, there will be an array for each type of entities, but the for the media entities, the array will not be present in the JSON if there are no media. This will force the array in the TwitterStatusMessageEntities class to be NULL, which isn't ideal from a developer point of view. This commit forces the array to always be an instance, so developers won't have to check for a NULL reference - and my own code won't throw up yellow error messages.