Facebook: Rewrote the code to explicitly use v1.0 of the API
Facebook has introduced a v2.0 of their API, while v1.0 can still be used until April 30th, 2015. When not explicitly specifying a version, the used version will depend on the creation date of the app. As a result of this Skybrud.Social has now been updated to explicitly use v1.0 of the API since all of the Facebook logic has been written for this version. I currently don't have a plan for when v2.0 will be supported.