Updated the PreBuildTool
Since I have a lot of builds between each release, and some of these builds are being used in internal projects, I use a small tool during prebuild to update the file version. The old version of this tool has some issues if the path to the solution contains any spaces, so this has now been fixed. The file version is based on the start date of the project, and this date is now specified as an argument rather than being hardcoded in the tool.