**Breaking change** Analytics: Rewrote some of the logic for fetching Analytics data for a given profile
Previously cell values were stored as string values (as they were received from the Analytics API), and the values for each cell were stored in a string array inside the AnalyticsDataRow class. This array has now been changed to hold instances of AnalyticsDataCell (this is a breaking change). Also new methods to convert the value to a desired type (eg. integer or double) has been added to both AnalyticsDataRow and AnalyticsDataCell. Besides holding the cell value, AnalyticsDataCell also has a reference to the parent AnalyticsDataColumnHeader so developers won't have to lookup this information manually.