Added reference to the JSON.net NuGet package (Newtonsoft.Json v6.0.5)
For classes in Skybrud.Social implementing or inheriting the JsonObject property, if an instance of a class was serialized using JSON.net, the generated JSON would also contain JsonObject, which isn't ideal since this is mostly an internal property. I have therefore now added a reference to the JSON.net package, and added the JsonIgnore attribute to the JsonObject property in class SocialJsonObject.

Skybrud.Social will however not be released with a dependency to the JSON.net package, the the changes only will have effect if you're already using JSON.net.