*Breaking change* Instagram: Added an options object for the Search method in the Locations endpoint
The Search method didn't allow developers to specify any of the extra options/parameters, so logic for this has now been added. Also renamed class InstagramLocationSearchOptions to InstagramRecentMediaSearchOptions, and introduced new class with old name InstagramLocationSearchOptions. This is a breaking change, but I rather want to make the change now than later.