Instagram: Added DoAuthenticatedGetRequest method to the InstagramOAuthClient (and updated the Locations endpoint to use this method)
The DoAuthenticatedGetRequest method will append the access token OR the client ID to the query string based on what has been specified for the InstagramOAuthClient. This is done to better support endpoint methods that support authentication via either an access token or a client ID. Currently only the Locations endpoint has been updated to use this new method, but some endpoints have similar logic internally. The plan is to update all endpoints to use the new method for consistency.