GitHub: Added method GetAuthenticatedGetRequest to class GitHubOAuthClient
GitHub seems to have changed their API since the my GitHub client was introduced. Apparently a user agent MUST be specified for all requests to the API. Not sure why though.

My plan with GetAuthenticatedGetRequest is that all GET requests to the API should use this method. An equivalent method should of course be implemented for all POSTS request when the time comes.