Facebook: Renamed class FacebookPostSummary to FacebookPost
The class was originally named FacebookPostSummary since it was used when fetching multiple posts and based on the idea than the response with multiple posts might not have all the properties as the response for a single post. I can't remember whether that was the case when the class was created, but it isn't now: the JSON for a post is identical in the two types of responses, so makes most sense to name the class FacebookPost.