RENAMING ALL THE CLASSES
Or something like that. Most of the classes in the HTTP logic now has the prefix "SocialHttp". Interfaces "IHttpQueryString" and "IHttpPostData" have also been introduced to representing a query string and POST data respectively.

For the OAuth 1.0a implementation, classes have now also been prefixed with "Social" - eg. "OAuthClient" has been renamed to "SocialOAuthClient". "SocialOAuthClient" now also inherits from "SocialHttpClient" since most of the logic for making HTTPS requests are the same. "SocialHttpClient" has a virtual "PrepareHttpRequest" method that "SocialOAuthClient" now overrides to update the request for the things specific to OAuth 1.0a (eg. the auhtorization header).

The changes in this commit aren't that heavily tested yet, but it seems to work like it should.