Added dependency for Skybrud.Social
Since some of the logic I have been adding recently can be used for other packages as well, the logic should instead be added to Skybrud.Essentials. Since I have started moving some of the logic to that package, Skybrud.Social should use the Skybrud.Essentials instead of having the same logic. For now, this means that Skybrud.Essentials should be a dependency.